rename halide2pencil.[cc|h]->halide_utils.[cc|h]
