+++ src/core/halide_utils.cc
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "tc/core/halide2pencil.h"
+#include "tc/core/halide_utils.h"
 
-#include <chrono>
+#include <map>
 #include <vector>
 
 #include "tc/core/flags.h"
